pyproject.toml
[project]
name = "anki-poems"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "genanki>=0.13.1",
    "python-slugify>=8.0.4",
    "requests>=2.32.4",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
